requirements.txt
# Core dependencies
httpx==0.27.2
playwright==1.48.0

# Database (latest versions compatible with Python 3.13)
supabase==2.9.0
postgrest==0.17.2

# LLM
openai==1.54.5

# Utilities
python-dotenv==1.0.1
aiofiles==24.1.0

# Logging and monitoring
colorlog==6.8.2
tqdm==4.67.1

